+++ src/textdecorator.ts
@@ -1,11 +1,12 @@
 import * as moment from 'moment';
+import * as ObjectPath from 'object-path';
 
 export class TextDecorator {
 
-    toTextView(commit: Object) : string {
+    static toTextView(commit: Object) : string {
         const dateNow = new Date();
         const author = commit['author'];
-        const dateText = this.toDateText(dateNow, new Date(author['timestamp'] * 1000));
+        const dateText = TextDecorator.toDateText(dateNow, new Date(author['timestamp'] * 1000));
 
         if (commit['hash'] === '0000000000000000000000000000000000000000') {
             return author['name'];
@@ -15,7 +16,7 @@ export class TextDecorator {
         }
     }
 
-    toDateText(dateNow: Date, dateThen: Date) : string {
+    static toDateText(dateNow: Date, dateThen: Date) : string {
 
         const momentNow = moment(dateNow);
         const momentThen = moment(dateThen);
@@ -40,5 +41,33 @@ export class TextDecorator {
         else {
             return months + ' months ago';
         }
+    }
+
+    static parseTokens(target: string, tokens: object = {}): string {
+        const tokenRegex = /\$\{([a-z\.\-]{1,})[,]*(|[a-z\-,]{1,})}/gi;
+
+        return target.replace(tokenRegex, (string: string, key: string, value: string): string => {
+            let currentToken = ObjectPath.get(tokens, key)
+
+            if (typeof currentToken === 'string') {
+                return currentToken;
+            }
+            else if (typeof currentToken === 'number') {
+                return currentToken.toString();
+            }
+            else if (typeof currentToken === 'function') {
+                let values = value.split(',');
+                let newString = currentToken.call(this, key, values);
+
+                if (typeof newString === 'string') {
+                    return newString;
+                }
+                else if (typeof newString === 'number') {
+                    return newString.toString();
+                }
+            }
+
+            return key;
+        });
     }
 }