+++ src/view.ts
@@ -1,24 +1,17 @@
-
 import {StatusBarItem} from 'vscode';
-export interface IView {
-    /**
-     * Refresh the view.
-     */
-    refresh(text: string): void;
-}
 
-export class StatusBarView implements IView {
+export class StatusBarView {
 
     private _statusBarItem: StatusBarItem;
 
     constructor(statusBarItem: StatusBarItem) {
         this._statusBarItem = statusBarItem;
-        this._statusBarItem.command = "extension.blame";
     }
 
-    refresh(text: string) {
+    refresh(text: string, hasCommand: boolean = true) {
         this._statusBarItem.text = '$(git-commit) ' + text;
-        this._statusBarItem.tooltip = 'git blame';
+        this._statusBarItem.tooltip = hasCommand ? 'git blame' : 'git blame - No info about current line';
+        this._statusBarItem.command = hasCommand ? "extension.blame" : undefined;
         this._statusBarItem.show();
     }
 }